Fitter: fix for nested dicts in nevergrad backend
Previously, I assumed that nested dictionaries were supported by
nevergrad. And to an extent they worked, but with more complicated
structures I was starting to see some problems were the initial_params
had different keys from the optimal params.
With this commit, the Nevergrad backend (like the scipy one)
also internally flattens the parameter dict and thus seems to be able
to handle more complicated dictionaries.

The unit tests were not complex enough to capture this, so I have added
a test case with a more complex parameter dict and bounds.