ase_objective_function: removed divide_by_n_atoms
this argument had too much overlap with the more general weight_cb
(literally being just a special case of it), but most of all the
implementation was misleading since for example for the energy obj.
function the correct normalization would have been 1/n_atoms^2 (square).

For these reasons I decided to remove it alltogether. Which sadly is a
breaking change. But, oh well, in these early stages it shouldnt really
matter too much.