combined_objective_function: rework
Removed the `gather_callback_function` in favour of `gather_meta_data`.
The main reason is that I don't want to worry about serializing any
potential callbacks later when I use this with the MPI wrapper.

Also it would be trivial to implement the gather_callback_function for
a user, but maybe it's not something we want to have to support in all
contexts.