Fitter: restructures fit_nevergrad
Now explicitly uses ask/tell interface and checks if the
recommendation.loss has been written.

Previously, it was always using the recommendation.loss to provide the
optimal loss. But that was ignoring that the loss is optional and thus
would sometimes give "None", causing parts of the code to crash.