REMOVED: MultiEnergyObjectiveFunction
since the combined objective function can now gather meta data
transparently the need for the MultiEnergyObjectiveFunction has
disappeared. It was more of a crutch anyways.

In the future everything should work on the basis of `meta_data` and
thus be transferrable between all types of objective functions with
the right interface.