ENH: huge refactor of abstract interface
ObjectiveFunctor
1. get_meta_data is gone
2. evaluate is gone
3. __call__ now takes on *optional* ctx

QuantitiyComputer: same-ish
1. get_meta_data is gone
2. evaluate is gone
3. __call__ now takes on *optional* ctx

The main rationale for these two changes was to simplify the code a bit.
Also I started to feel like the `get_meta_data` with caching was just
legacy and holding us back. So I axed it.
Now everything should be super transparent regarding the contexts.

Had to change the code in a lot of places due to the above.

Then, I also touched the combined objective function.
The functionality of `gather_meta_data` now lives in the ctx.
And there is an additional "gather" step after __call__.

The two wrappers async and MPI had to be changed accordingly.

Then I also introduced a convenient module to turn plain old functions
into QuantitiyComputers and ObjectiveFunctors respectively.