abstract_objective_function: formalize child_ctx
EvaluateContexts can now spawn a child context,
which is tracked internally and whose meta data
can be collected.

Crucially this child context gets a shallow copy
of the temp data of the context.

This is mainly beneficial because it creates a
mechanism for the context of a
combined_objective_function to somewhat propagate
to the contexts of the terms ("the cildren").

This makes is possible to define a compute grid of
nodes for the terms of the objective function by
injecting data into the contexts on the call site.

To further enable this use case, the fitter can
now use pre-created contexts, so that a user may
perform such a setup before fitting.

Further, the combined objective function now
supports a `context_modifier` protocol, which
makes it possible to modify the child contexts
after creation.