abstract_objective_function: changed spawn_child
It is now "spawn_children" and spawns multiple
children at once, while *overwriting* the current
children.

The main reason for this change was that it was
easy to have a lot of "dead" child context if
one keeps a parent context alive and makes
multiple calls to a combined objective function.

The old `spawn_child` function would just append
to the internal _children list of the parent
context and it would keep growing and growing.

This new API is a bit harder to misuse I hope.